Fix: The repo of annie now contains 'v' in its tag
The repository of annie now contains a letter 'v' in its tag in new
release "v0.11.0". None of the previous tags contained the letter
'v', such as "0.10.2", "0.9.8".
Anyone trying to upgrade annie via `Deploy.bat` will encounter an
error that failing to download the new release because of a wrong URL.